common/mlx5: reduce HCA attribute type sizes
A number of `log_max_*` fields' types are unnecessarily large, and can
be reduced to `uint8_t`.

Acked-by: Viacheslav Ovsiienko <viacheslavo@nvidia.com>

+++ drivers/common/mlx5/mlx5_devx_cmds.h
@@ -268,10 +268,10 @@ struct mlx5_hca_attr {
 	struct mlx5_hca_flex_attr flex;
 	struct mlx5_hca_crypto_mmo_attr crypto_mmo;
 	uint8_t log_max_wq_sz;
-	int log_max_qp_sz;
-	int log_max_cq_sz;
-	int log_max_qp;
-	int log_max_cq;
+	uint8_t log_max_qp_sz;
+	uint8_t log_max_cq_sz;
+	uint8_t log_max_qp;
+	uint8_t log_max_cq;
 	uint32_t log_max_pd;
 	uint32_t log_max_mrw_sz;
 	uint32_t log_max_srq;
